test: add unit test for "write" file
All functions now have functions for unit test.
There is two variables for "samplePath", & "sampleMessage". These
contains sample values which could change in each tests, but changes
doesn't make error by this global variables.
How does them work:
The unit will call the function & log the return value. Then check that
the wanted file is exist or not. At the end it will handle the error and
log the result. If the result was succesful, the sample file will
remove.