feat: create `execText`
This text will help us to build query for inserting data to
`FileMetadataTable`. The query has a table name(which is a global
variable), and each input parameter after each other. It is also in
order of its table attributes order.
The null value will endicate the `id` which is an auto increamental
attribute.