feat: add function to insert data to `AnalyzeFileResultTable`
This function will insert data to the mentioned table. It takes
`fileMetadataId` which is a foreign key to the `FileMetadataTable.id`.
It also takes `language` which is a string(that indicates file
language). On the go, it also takes `codeSize`, `commentSize`,
`blankLines`, & `total` that are int(They are size). It also returns
error for SAFTY.