feat: create `execText`
This text will help us to build query for inserting data to
`AnalyzeFileResult`. The query has a table name(which is a global
variable). A `fileMetadataId` which is a foreign key to
`FileMetadataTable.id`. Down the road, there are the `language`,
`codeSize`, `commentSize`, `blankSize`, & `total` that we took in
function parameters.
The null value will endicate this table `id` which is an auto
increamental attribute.