feat: select all data in `TblFileMetadata`
By calling the `Query` function we will be able to get datas row by row
with a query string. The query string will select all attributs of
`TblFileMetadata` & all `TblAnalyzeFileResult` except The
`FileMetadataId` which is the same as `TblFileMetadata.Id`. Down the
road, it will join by the `TblAnalyzeFileResult` of their keys.
For SAFTY, The query function connection will close when the function
finished.
Error handling also added.