test[feat]: test the function
The function will call and the result will be save. Return error will be
handled with `assert` function.
For actual testing, we will compare an empty array of
`AnalyzeFileResult` with the array that we saved earlier. They should
not be equal unless we got error and the select query goes wrong.